refactored the server code
